WIP - get_work test
Note to include in PR:

get_work's overcome parameter will cause the following:

Given a chain of three blocks, with equal difficulty d and overcome
set to d as well, get_work will return 2d. This is because the
`out_work <= overcome` condition will be true after the first block is
considered and will continue to capture work for second block too. Is
this expected behavior?

One of the old commented out test for get_work seemed to suggest that
get_work should return d in the above setup. So I am not sure if there
is something wrong in the test setup.